requirements.txt
Django==4.0.1
djongo pytz pymongo==3.12.3
djangorestframework
mysqlclient
Pillow
requests